Struct uint_zigzag::Uint

pub struct Uint(pub u128);

Uint implements zig-zag encoding to represent integers as binary sequences

Tuple Fields
0: u128
Implementations

impl Uint

pub const MAX_BYTES: usize = 19usize

The maximum number of bytes a uint will consume

pub fn peek(value: &[u8]) -> Option<usize>

Peek returns the number of bytes that would be read or None if no Uint cannot be read

use uint_zigzag::Uint;

let buffer = [0x34u8];

let out = Uint::peek(&buffer);

assert!(out.is_some());

let out = Uint::peek(&[]);

assert!(out.is_none());

pub fn to_bytes<M: AsMut<[u8]>>(&self, buffer: M)

Zig-zag encoding, any length from 1 to MAX_BYTES into buffer buffer must be big enough to hold the result

 use uint_zigzag::Uint;

let mut buffer = [0u8, 0u8];
let mut u = Uint::from(127);
u.to_bytes(&mut buffer);

assert_eq!(buffer, [0x7Fu8, 0u8]);

u = Uint::from(128);
u.to_bytes(&mut buffer);
assert_eq!(buffer, [0x80u8, 1u8]);

pub fn to_bytes_with_length(self, buffer: &mut [u8]) -> usize

Same as to_bytes except it returns how many bytes were actually used

pub fn to_vec(&self) -> Vec<u8>

Zig-zag encoding, any length from 1 to MAX_BYTES

use uint_zigzag::Uint;

let u = Uint(345678);
let out = u.to_vec();

assert_eq!(out.as_slice(), &[206, 140, 21]);

let u = Uint(2048);
let out = u.to_vec();
assert_eq!(out.as_slice(), &[128, 16]);

pub fn to_writer<W: Write>(&self, writer: &mut W) -> Result<usize, Error>

Write bytes to a stream

pub fn from_reader<R: Read>(reader: &mut R) -> Result<Self, Error>

Read bytes from a stream